PatientInfo 2

+++ context/models/PatientDocument.cs
@@ -19,7 +19,7 @@ namespace Dentistry.Models
         public string Comment { get; set; }
         public bool IsDeleted { get; set; }
 
-        public long PatientId { get; set; }
+        public int PatientId { get; set; }
         public virtual Patient Patient { get; set; }
     }
 }
