+++ context/models/PatientTooth.cs
@@ -65,7 +65,7 @@ namespace Dentistry.Models
         public string Description { get; set; }
         public bool IsDeleted { get; set; }
 
-        public long PatientId { get; set; }
+        public int PatientId { get; set; }
         public virtual Patient Patient { get; set; }
 
         public int? ToothId { get; set; }
